游客模式

+++ cloud-base/base-core/src/main/java/pro/simplecloud/utils/Timer.java
@@ -15,8 +15,8 @@ public class Timer {
         start = System.currentTimeMillis();
     }
 
-    public int end() {
+    public long end() {
         long end = System.currentTimeMillis();
-        return (int) (end - start);
+        return end - start;
     }
 }
